[BUGFIX] Fixes for PR
